src/log-analytics-function/requirements.txt
azure-functions
azure-identity
azure-monitor-query>=2.0.0
